# docs/swagger/swagger-sdce-1.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "email" : "onap-discuss@lists.onap.org",
      "name" : "ONAP",
      "url" : "https://onap.readthedocs.io"
    },
    "description" : "SDC API for onboarding (SDCE-1)",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "title" : "SPC API: SDCE-1",
    "version" : "1.0"
  },
  "servers" : [ {
    "description" : "SDCE-1 APIs",
    "url" : "/sdc"
  } ],
  "paths" : {
...
    "/v1.0/vendor-license-models/{vlmId}/versions/{versionId}/entitlement-pools" : {
      "get" : {
        "description" : "List vendor entitlement pools",
        "operationId" : "listEntitlementPools",
        "parameters" : [ {
          "description" : "Vendor license model Id",
          "in" : "path",
          "name" : "vlmId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/EntitlementPoolEntityDto"
                  }
                }
              }
            }
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor License Model - Entitlement Pools" ]
      },
      "post" : {
        "description" : "Create vendor entitlement pool",
        "operationId" : "createEntitlementPool",
        "parameters" : [ {
          "description" : "Vendor license model Id",
          "in" : "path",
          "name" : "vlmId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EntitlementPoolRequestDto"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor License Model - Entitlement Pools" ]
      }
    },
    "/v1.0/vendor-license-models/{vlmId}/versions/{versionId}/entitlement-pools/{entitlementPoolId}" : {
      "delete" : {
        "description" : "Delete vendor entitlement pool",
        "operationId" : "deleteEntitlementPool",
        "parameters" : [ {
          "description" : "Vendor license model Id",
          "in" : "path",
          "name" : "vlmId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "entitlementPoolId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor License Model - Entitlement Pools" ]
      },
      "get" : {
        "description" : "Get vendor entitlement pool",
        "operationId" : "getEntitlementPool",
        "parameters" : [ {
          "description" : "Vendor license model Id",
          "in" : "path",
          "name" : "vlmId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "entitlementPoolId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntitlementPoolEntityDto"
                }
              }
            }
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor License Model - Entitlement Pools" ]
      },
      "put" : {
        "description" : "Update vendor entitlement pool",
        "operationId" : "updateEntitlementPool",
        "parameters" : [ {
          "description" : "Vendor license model Id",
          "in" : "path",
          "name" : "vlmId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "entitlementPoolId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EntitlementPoolRequestDto"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor License Model - Entitlement Pools" ]
      }
    },
    "/v1.0/vendor-license-models/{vlmId}/versions/{versionId}/entitlement-pools/{entitlementPoolId}/limits" : {
      "get" : {
        "description" : "List vendor entitlement pool limits",
        "operationId" : "listLimits",
        "parameters" : [ {
          "description" : "Vendor license model Id",
          "in" : "path",
          "name" : "vlmId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model Entitlement Pool Id",
          "in" : "path",
          "name" : "entitlementPoolId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/LimitRequestDto"
                  }
                }
              }
            }
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor License Model - Entitlement Pool Limits" ]
      },
      "post" : {
        "description" : "Create vendor entitlement pool limits",
        "operationId" : "createLimit",
        "parameters" : [ {
          "description" : "Vendor license model Id",
          "in" : "path",
          "name" : "vlmId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model Entitlement Pool Id",
          "in" : "path",
          "name" : "entitlementPoolId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LimitRequestDto"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor License Model - Entitlement Pool Limits" ]
      }
    },
    "/v1.0/vendor-license-models/{vlmId}/versions/{versionId}/entitlement-pools/{entitlementPoolId}/limits/{limitId}" : {
      "delete" : {
        "description" : "Delete vendor entitlement pool limit",
        "operationId" : "deleteLimit",
        "parameters" : [ {
          "description" : "Vendor license model Id",
          "in" : "path",
          "name" : "vlmId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model Entitlement pool Id",
          "in" : "path",
          "name" : "entitlementPoolId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "limitId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor License Model - Entitlement Pool Limits" ]
      },
      "get" : {
        "description" : "Get vendor entitlement pool limit",
        "operationId" : "getLimit",
        "parameters" : [ {
          "description" : "Vendor license model Id",
          "in" : "path",
          "name" : "vlmId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model Entitlement Pool Id",
          "in" : "path",
          "name" : "entitlementPoolId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model Entitlement Pool Limit Id",
          "in" : "path",
          "name" : "limitId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LimitEntityDto"
                }
              }
            }
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor License Model - Entitlement Pool Limits" ]
      },
      "put" : {
        "description" : "Update vendor entitlement pool limit",
        "operationId" : "updateLimit",
        "parameters" : [ {
          "description" : "Vendor license model Id",
          "in" : "path",
          "name" : "vlmId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model Entitlement Pool Id",
          "in" : "path",
          "name" : "entitlementPoolId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "limitId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LimitRequestDto"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor License Model - Entitlement Pool Limits" ]
      }
    },
...
    "/v1.0/vendor-license-models/{vlmId}/versions/{versionId}/license-key-groups/{licenseKeyGroupId}/limits" : {
      "get" : {
        "description" : "List vendor license key group limits",
        "operationId" : "listLimits_1",
        "parameters" : [ {
          "description" : "Vendor license model Id",
          "in" : "path",
          "name" : "vlmId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model License Key Group Id",
          "in" : "path",
          "name" : "licenseKeyGroupId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/LimitEntityDto"
                  }
                }
              }
            }
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor License Model - License Key Group Limits" ]
      },
      "post" : {
        "description" : "Create vendor license key group limit",
        "operationId" : "createLimit_1",
        "parameters" : [ {
          "description" : "Vendor license model Id",
          "in" : "path",
          "name" : "vlmId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model License Key Group Id",
          "in" : "path",
          "name" : "licenseKeyGroupId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LimitRequestDto"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor License Model - License Key Group Limits" ]
      }
    },
    "/v1.0/vendor-license-models/{vlmId}/versions/{versionId}/license-key-groups/{licenseKeyGroupId}/limits/{limitId}" : {
      "delete" : {
        "description" : "Delete vendor license key group limit",
        "operationId" : "deleteLimit_1",
        "parameters" : [ {
          "description" : "Vendor license model Id",
          "in" : "path",
          "name" : "vlmId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model license key group Id",
          "in" : "path",
          "name" : "licenseKeyGroupId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "limitId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor License Model - License Key Group Limits" ]
      },
      "get" : {
        "description" : "Get vendor entitlement pool limit",
        "operationId" : "getLimit_1",
        "parameters" : [ {
          "description" : "Vendor license model Id",
          "in" : "path",
          "name" : "vlmId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model License Key Group",
          "in" : "path",
          "name" : "licenseKeyGroupId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model License Key Group Limit Id",
          "in" : "path",
          "name" : "limitId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LimitEntityDto"
                }
              }
            }
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor License Model - License Key Group Limits" ]
      },
      "put" : {
        "description" : "Update vendor license key group limit",
        "operationId" : "updateLimit_1",
        "parameters" : [ {
          "description" : "Vendor license model Id",
          "in" : "path",
          "name" : "vlmId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor license model License Key Group Id",
          "in" : "path",
          "name" : "licenseKeyGroupId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "limitId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LimitRequestDto"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor License Model - License Key Group Limits" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "EntitlementPoolEntityDto" : {
        "type" : "object",
        "description" : "EntitlementPoolEntity",
        "properties" : {
          "description" : {
            "type" : "string",
            "maxLength" : 1000,
            "minLength" : 0
          },
          "expiryDate" : {
            "type" : "string"
          },
          "getversionUUID" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "increments" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          },
          "manufacturerReferenceNumber" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "name" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          },
          "operationalScope" : {
            "$ref" : "#/components/schemas/MultiChoiceOrOtherDtoOperationalScope"
          },
          "referencingFeatureGroups" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          },
          "startDate" : {
            "type" : "string"
          },
          "thresholdUnits" : {
            "type" : "string",
            "enum" : [ "Absolute", "Percentage" ]
          },
          "thresholdValue" : {
            "type" : "integer",
            "format" : "int32"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "Universal", "Unique", "One_Time" ]
          }
        },
        "required" : [ "manufacturerReferenceNumber", "name", "type" ]
      },
      "EntitlementPoolRequestDto" : {
        "type" : "object",
        "description" : "EntitlementPoolRequest",
        "properties" : {
          "description" : {
            "type" : "string",
            "maxLength" : 1000,
            "minLength" : 0
          },
          "expiryDate" : {
            "type" : "string"
          },
          "increments" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          },
          "manufacturerReferenceNumber" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "name" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          },
          "operationalScope" : {
            "$ref" : "#/components/schemas/MultiChoiceOrOtherDtoOperationalScope"
          },
          "startDate" : {
            "type" : "string"
          },
          "thresholdUnits" : {
            "type" : "string",
            "enum" : [ "Absolute", "Percentage" ]
          },
          "thresholdValue" : {
            "type" : "integer",
            "format" : "int32"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "Universal", "Unique", "One_Time" ]
          }
        },
        "required" : [ "manufacturerReferenceNumber", "name", "type" ]
      },
...
      "LimitEntityDto" : {
        "type" : "object",
        "properties" : {
          "aggregationFunction" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "metric" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "time" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "unit" : {
            "type" : "string"
          },
          "value" : {
            "type" : "string"
          }
        }
      },
      "LimitRequestDto" : {
        "type" : "object",
        "description" : "LimitRequest",
        "properties" : {
          "aggregationFunction" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string",
            "maxLength" : 1000,
            "minLength" : 0
          },
          "metric" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          },
          "time" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "unit" : {
            "type" : "string"
          },
          "value" : {
            "type" : "string"
          }
        },
        "required" : [ "metric", "name", "type", "value" ]
      },
...
      "MultiChoiceOrOtherDtoOperationalScope" : {
        "type" : "object",
        "properties" : {
          "choices" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "Network_Wide", "Availability_Zone", "Data_Center", "Tenant", "VM", "CPU", "Core", "Other" ]
            },
            "maxItems" : 2147483647,
            "minItems" : 1,
            "uniqueItems" : true
          },
          "other" : {
            "type" : "string"
          }
        },
        "required" : [ "choices", "other" ]
      },
...
    }
  }
}